sched: log the cpu utilization at rq
The cpu's utilization is to measure how busy is the cpu.
        util = cpu_rq(cpu)->avg.runnable_avg_sum
                / cpu_rq(cpu)->avg.runnable_avg_period;

Since the util is no more than 1, we use its percentage value in later
caculations. And set the the FULL_UTIL as 99%.

In later power aware scheduling, we are sensitive for how busy of the
cpu, not how weight of its load. As to power consuming, it is more
related with busy time, not the load weight.

Signed-off-by: Alex Shi <alex.shi@intel.com>
adapted for Mako kernel from LKML

Signed-off-by: faux123 <reioux@gmail.com>